requirements.txt
accelerate==1.12.0
anyio==4.12.1
bitsandbytes==0.49.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
filelock==3.20.3
Flask==3.1.2
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
mlx==0.30.1
mlx-lm==0.30.2
mlx-metal==0.30.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pillow==12.1.0
protobuf==6.33.3
psutil==7.2.1
pymongo==4.16.0
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
shellingham==1.5.4
slack_bolt==1.27.0
slack_sdk==3.39.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==5.0.0rc1
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
pytest==8.3.4